requirements.txt
django==6.0.4
django-environ==0.13.0
psycopg2-binary==2.9.12
pillow==12.2.0
django-storages==1.14.6
boto3==1.38.3
django-phonenumber-field==8.4.0
phonenumbers==9.0.28
yookassa==3.10.1
django-widget-tweaks==1.5.1
